Class Service has been moved to Framework\DI namespace.
