+++ framework/Application.php
@@ -14,6 +14,8 @@ use Framework\Response\ResponseRedirect;
 use Framework\Exception\BadResponseTypeException;
 use Framework\Exception\HttpNotFoundException;
 use Framework\Exception\AuthRequredException;
+use Framework\DI\Service;
+
 
 
 /**
